generalise create network method
